+++ mailspy-frontend/src/inversify.config.ts
@@ -28,6 +28,7 @@ import { HttpService } from "services/http/HttpService";
 import { MailParserService } from "services/mail/MailParserService";
 import { HtmlService } from "services/html/HtmlService";
 import { FileDownloadService } from "services/download/FileDownloadService";
+import { AttachmentIconService } from "services/mail/AttachmentIconService";
 
 const container = new Container({
     defaultScope: "Singleton"
@@ -38,6 +39,7 @@ container.bind<HttpService>(HttpService).toSelf();
 container.bind<MailParserService>(MailParserService).toSelf();
 container.bind<HtmlService>(HtmlService).toSelf();
 container.bind<FileDownloadService>(FileDownloadService).toSelf();
+container.bind<AttachmentIconService>(AttachmentIconService).toSelf();
 
 container.bind<DOMParser>(DOMParser).toConstantValue(new DOMParser());
 container.bind<XMLSerializer>(XMLSerializer).toConstantValue(new XMLSerializer());